Use uint64_t for hash values
They are uint64_t. Usually size_t is a uint64_t as
well, but not always (e.g. Win32).

+++ openvpn/addr/ip.hpp
@@ -1022,7 +1022,7 @@ class Addr
     }
 
 #ifdef USE_OPENVPN_HASH
-    std::size_t hashval() const
+    std::uint64_t hashval() const
     {
         Hash64 h;
         hash(h);
