+++ openvpn/ssl/sess_ticket.hpp
@@ -104,7 +104,7 @@ class TLSSessionTicketBase
         }
 
 #ifdef USE_OPENVPN_HASH
-        std::size_t hashval() const
+        std::uint64_t hashval() const
         {
             Hash64 h;
             hash(h);
